% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.kfold.R
\name{accuracy.kfold}
\alias{accuracy.kfold}
\title{accuracy.kfold}
\usage{
accuracy.kfold(df.VAR, k = 5, n_ahead, lags, var_index = 1)
}
\arguments{
\item{df.VAR}{data.frame : data of the VAR (without dates)}

\item{k}{numeric : number of fold cross validations}

\item{n_ahead}{numeric :horizon of forecast}

\item{lags}{numeric : lags of the VAR}

\item{var_index}{numeric : index of the variable to forecast}
}
\value{
a list with fit measures for each fold
}
\description{
This function it splits the data in k training and test datasets and returns measures of fit for each fold.
}
